requirements.txt
streamlit==1.31.1
ultralytics==8.0.207
opencv-python-headless==4.8.1.78
Pillow==9.5.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0